Moved stuff around, fixed some api errors
