+++ group_project_v2/group_project.py
@@ -5,16 +5,12 @@
 from collections import OrderedDict
 
 import logging
-import json
 import itertools
 from lazy.lazy import lazy
 from opaque_keys import InvalidKeyError
 from opaque_keys.edx.locator import BlockUsageLocator
-import pytz
-import webob
-from datetime import datetime, timedelta
 
-from django.utils import html
+
 from django.utils.translation import ugettext as _
 
 from xblock.core import XBlock
@@ -25,22 +21,19 @@ from xblock.validation import ValidationMessage
 
 from xblockutils.studio_editable import StudioEditableXBlockMixin, StudioContainerXBlockMixin
 
-from group_project_v2.mixins import ChildrenNavigationXBlockMixin, UserAwareXBlockMixin, CourseAwareXBlockMixin, \
+from group_project_v2.mixins import (
+    ChildrenNavigationXBlockMixin, UserAwareXBlockMixin, CourseAwareXBlockMixin,
     WorkgroupAwareXBlockMixin, XBlockWithComponentsMixin, XBlockWithPreviewMixin
+)
+from group_project_v2.notifications import ActivityNotificationsMixin
 from group_project_v2.project_navigator import GroupProjectNavigatorXBlock
-from group_project_v2.utils import loader, OutsiderDisallowedError, make_key, outsider_disallowed_protected_view
+from group_project_v2.utils import loader, make_key, outsider_disallowed_protected_view
 from group_project_v2.stage import (
     BasicStage, SubmissionStage, PeerReviewStage, GroupReviewStage,
     PeerAssessmentStage, GroupAssessmentStage, CompletionStage
 )
 from group_project_v2.project_api import ProjectAPIXBlockMixin
 from group_project_v2.api_error import ApiError
-
-try:
-    from edx_notifications.data import NotificationMessage  # pylint: disable=import-error
-except ImportError:
-    # Notifications is an optional runtime configuration, so it may not be available for import
-    pass
 
 
 log = logging.getLogger(__name__)  # pylint: disable=invalid-name
@@ -120,69 +113,14 @@ class GroupProjectXBlock(
         return validation
 
 
-class ActivityNavigationViewMixin(object):
-    def navigation_view(self, context):
-        fragment = Fragment()
-
-        stage_contents = []
-        for child in self.stages:
-            child_fragment = child.render('navigation_view', context)
-            fragment.add_frag_resources(child_fragment)
-            stage_contents.append(child_fragment.content)
-
-        context = {'activity': self, 'stage_contents': stage_contents}
-        fragment.add_content(loader.render_template("templates/html/activity/navigation_view.html", context))
-
-        return fragment
-
-
-class ActivityResourcesViewMixin(object):
-    def resources_view(self, context):
-        fragment = Fragment()
-
-        has_resources = any([bool(stage.resources) for stage in self.stages])
-
-        stage_contents = []
-        for child in self.stages:
-            child_fragment = child.render('resources_view', context)
-            fragment.add_frag_resources(child_fragment)
-            stage_contents.append(child_fragment.content)
-
-        context = {'activity': self, 'stage_contents': stage_contents, 'has_resources': has_resources}
-        fragment.add_content(loader.render_template("templates/html/activity/resources_view.html", context))
-
-        return fragment
-
-
-class ActivitySubmissionsViewMixin(object):
-    def submissions_view(self, context):
-        fragment = Fragment()
-
-        target_stages = [stage for stage in self.stages if isinstance(stage, SubmissionStage)]
-
-        has_submissions = any([stage.has_submissions for stage in target_stages])
-
-        stage_contents = []
-        for child in target_stages:
-            child_fragment = child.render('submissions_view', context)
-            fragment.add_frag_resources(child_fragment)
-            stage_contents.append(child_fragment.content)
-
-        context = {'activity': self, 'stage_contents': stage_contents, 'has_submissions': has_submissions}
-        fragment.add_content(loader.render_template("templates/html/activity/submissions_view.html", context))
-
-        return fragment
-
-
 # TODO: enable and fix these violations
 # pylint: disable=unused-argument,invalid-name
 @XBlock.wants('notifications')
 @XBlock.wants('courseware_parent_info')
 class GroupActivityXBlock(
-    XBlockWithPreviewMixin, XBlockWithComponentsMixin, ProjectAPIXBlockMixin,
+    XBlockWithPreviewMixin, XBlockWithComponentsMixin, ProjectAPIXBlockMixin, ActivityNotificationsMixin,
     XBlock, StudioEditableXBlockMixin, StudioContainerXBlockMixin,
-    ChildrenNavigationXBlockMixin, CourseAwareXBlockMixin, UserAwareXBlockMixin, WorkgroupAwareXBlockMixin,
-    ActivityNavigationViewMixin, ActivityResourcesViewMixin, ActivitySubmissionsViewMixin
+    ChildrenNavigationXBlockMixin, CourseAwareXBlockMixin, UserAwareXBlockMixin, WorkgroupAwareXBlockMixin
 ):
     """
     XBlock providing a group activity project for a group of students to collaborate upon
@@ -225,10 +163,6 @@ class GroupActivityXBlock(
     @property
     def id(self):
         return self.scope_ids.usage_id
-
-    @property
-    def is_group_member(self):
-        return self.user_id in [u["id"] for u in self.workgroup["users"]]
 
     @property
     def is_admin_grader(self):
@@ -278,36 +212,6 @@ class GroupActivityXBlock(
 
         return self.stages[0] if self.stages else None
 
-    @property
-    def team_members(self):
-        """
-        Returns teammates to review. May throw `class`: OutsiderDisallowedError
-        """
-        if not self.is_group_member:
-            return []
-
-        try:
-            return [
-                self.project_api.get_user_details(team_member["id"])
-                for team_member in self.workgroup["users"]
-                if self.user_id != int(team_member["id"])
-            ]
-        except ApiError:
-            return []
-
-    @property
-    def review_groups(self):
-        """
-        Returns groups to review. May throw `class`: OutsiderDisallowedError
-        """
-        if not self.is_group_member:
-            return self.workgroup
-
-        try:
-            return self.project_api.get_workgroups_to_review(self.user_id, self.course_id, self.content_id)
-        except ApiError:
-            return []
-
     @outsider_disallowed_protected_view
     def student_view(self, context):
         """
@@ -333,6 +237,83 @@ class GroupActivityXBlock(
 
         return fragment
 
+    def navigation_view(self, context):
+        fragment = Fragment()
+
+        stage_contents = []
+        for child in self.stages:
+            child_fragment = child.render('navigation_view', context)
+            fragment.add_frag_resources(child_fragment)
+            stage_contents.append(child_fragment.content)
+
+        context = {'activity': self, 'stage_contents': stage_contents}
+        fragment.add_content(loader.render_template("templates/html/activity/navigation_view.html", context))
+
+        return fragment
+
+    def resources_view(self, context):
+        fragment = Fragment()
+
+        has_resources = any([bool(stage.resources) for stage in self.stages])
+
+        stage_contents = []
+        for child in self.stages:
+            child_fragment = child.render('resources_view', context)
+            fragment.add_frag_resources(child_fragment)
+            stage_contents.append(child_fragment.content)
+
+        context = {'activity': self, 'stage_contents': stage_contents, 'has_resources': has_resources}
+        fragment.add_content(loader.render_template("templates/html/activity/resources_view.html", context))
+
+        return fragment
+
+    def submissions_view(self, context):
+        fragment = Fragment()
+
+        target_stages = [stage for stage in self.stages if isinstance(stage, SubmissionStage)]
+
+        has_submissions = any([stage.has_submissions for stage in target_stages])
+
+        stage_contents = []
+        for child in target_stages:
+            child_fragment = child.render('submissions_view', context)
+            fragment.add_frag_resources(child_fragment)
+            stage_contents.append(child_fragment.content)
+
+        context = {'activity': self, 'stage_contents': stage_contents, 'has_submissions': has_submissions}
+        fragment.add_content(loader.render_template("templates/html/activity/submissions_view.html", context))
+
+        return fragment
+
+    def mark_complete(self, user_id):
+        try:
+            self.project_api.mark_as_complete(self.course_id, self.content_id, user_id)
+        except ApiError as e:
+            # 409 indicates that the completion record already existed. That's ok in this case
+            if e.code != 409:
+                raise
+
+    def validate_field_data(self, validation, data):
+        super(GroupActivityXBlock, self).validate_field_data(validation, data)
+        should_be_ints = ('weight', 'group_reviews_required_count', 'user_review_count')
+        for field_name in should_be_ints:
+            try:
+                int(getattr(data, field_name))
+            except (TypeError, ValueError):
+                message = _(u"{field_name} must be integer, {field_value} given").format(
+                    field_name=field_name, field_value=getattr(data, field_name)
+                )
+                validation.add(ValidationMessage(ValidationMessage.ERROR, message))
+
+    def calculate_and_send_grade(self, group_id):
+        grade_value = self.calculate_grade(group_id)
+        if grade_value:
+            self.assign_grade_to_group(group_id, grade_value)
+
+            workgroup = self.project_api.get_workgroup_by_id(group_id)
+            for u in workgroup["users"]:
+                self.mark_complete(u["id"])
+
     def assign_grade_to_group(self, group_id, grade_value):
         self.project_api.set_group_grade(
             group_id,
@@ -351,18 +332,13 @@ class GroupActivityXBlock(
                 "content_id": self.content_id,
             }
         )
+        self.runtime.publish(self, 'grade', {
+            'value': grade_value,
+            'max_value': self.weight,
+        })
         notifications_service = self.runtime.service(self, 'notifications')
         if notifications_service:
             self.fire_grades_posted_notification(group_id, notifications_service)
-
-    def calculate_and_send_grade(self, group_id):
-        grade_value = self.calculate_grade(group_id)
-        if grade_value:
-            self.assign_grade_to_group(group_id, grade_value)
-
-            workgroup = self.project_api.get_workgroup_by_id(group_id)
-            for u in workgroup["users"]:
-                self.mark_complete(u["id"])
 
     def calculate_grade(self, group_id):
 
@@ -434,324 +410,3 @@ class GroupActivityXBlock(
         group_grade = round(mean(reviewer_grades)) if len(reviewer_grades) > 0 else None
 
         return group_grade
-
-    def mark_complete(self, user_id):
-        try:
-            self.project_api.mark_as_complete(self.course_id, self.content_id, user_id)
-        except ApiError as e:
-            # 409 indicates that the completion record already existed. That's ok in this case
-            if e.code != 409:
-                raise
-
-    def validate_field_data(self, validation, data):
-        super(GroupActivityXBlock, self).validate_field_data(validation, data)
-        should_be_ints = ('weight', 'group_reviews_required_count', 'user_review_count')
-        for field_name in should_be_ints:
-            try:
-                int(getattr(data, field_name))
-            except (TypeError, ValueError):
-                message = _(u"{field_name} must be integer, {field_value} given").format(
-                    field_name=field_name, field_value=getattr(data, field_name)
-                )
-                validation.add(ValidationMessage(ValidationMessage.ERROR, message))
-
-    def get_courseware_info(self, courseware_parent_info_service):
-        activity_name = self.display_name
-        activity_location = None
-        stage_name = self.display_name
-        stage_location = None
-        project_name = None
-        project_location = None
-
-        try:
-            if courseware_parent_info_service:
-                # First get Unit (first parent)
-                stage_info = courseware_parent_info_service.get_parent_info(
-                    self.location
-                )
-                stage_location = stage_info['location']
-                stage_name = stage_info['display_name']
-
-                # Then get Sequence (second parent)
-                activity_courseware_info = courseware_parent_info_service.get_parent_info(
-                    stage_location
-                )
-                activity_name = activity_courseware_info['display_name']
-                activity_location = activity_courseware_info['location']
-
-                project_courseware_info = courseware_parent_info_service.get_parent_info(
-                    activity_location
-                )
-                project_name = project_courseware_info['display_name']
-                project_location = project_courseware_info['location']
-
-        except Exception, ex:  # pylint: disable=broad-except
-            # Can't look this up then log and just use the default
-            # which is our display_name
-            log.exception(ex)
-
-        return {
-            'stage_name': stage_name,
-            'stage_location': stage_location,
-            'activity_name': activity_name,
-            'activity_location': activity_location,
-            'project_name': project_name,
-            'project_location': project_location,
-        }
-
-    # TODO: this fire_* methods are mostly identical - might make sense to refactor into single method
-    def fire_file_upload_notification(self, notifications_service):
-        try:
-            # this NotificationType is registered in the list of default Open edX Notifications
-            msg_type = notifications_service.get_notification_type('open-edx.xblock.group-project.file-uploaded')
-
-            workgroup_user_ids = []
-            uploader_username = ''
-            for user in self.workgroup['users']:
-                # don't send to ourselves
-                if user['id'] != self.user_id:
-                    workgroup_user_ids.append(user['id'])
-                else:
-                    uploader_username = user['username']
-
-            # get the activity name which is simply our hosting
-            # Sequence's Display Name, so call out to a new xBlock
-            # runtime Service
-
-            courseware_info = self.get_courseware_info(self.runtime.service(self, 'courseware_parent_info'))
-
-            activity_name = courseware_info['activity_name']
-            activity_location = courseware_info['activity_location']
-
-            msg = NotificationMessage(
-                msg_type=msg_type,
-                namespace=unicode(self.course_id),
-                payload={
-                    '_schema_version': 1,
-                    'action_username': uploader_username,
-                    'activity_name': activity_name,
-                }
-            )
-
-            #
-            # add in all the context parameters we'll need to
-            # generate a URL back to the website that will
-            # present the new course announcement
-            #
-            # IMPORTANT: This can be changed to msg.add_click_link() if we
-            # have a particular URL that we wish to use. In the initial use case,
-            # we need to make the link point to a different front end website
-            # so we need to resolve these links at dispatch time
-            #
-            msg.add_click_link_params({
-                'course_id': unicode(self.course_id),
-                'activity_location': unicode(activity_location) if activity_location else '',
-            })
-
-            # NOTE: We're not using Celery here since we are expectating that we
-            # will have only a very small handful of workgroup users
-            notifications_service.bulk_publish_notification_to_users(
-                workgroup_user_ids,
-                msg
-            )
-        except Exception, ex:  # pylint: disable=broad-except
-            # While we *should* send notification, if there is some
-            # error here, we don't want to blow the whole thing up.
-            # So log it and continue....
-            log.exception(ex)
-
-    def fire_grades_posted_notification(self, group_id, notifications_service):
-        try:
-            # this NotificationType is registered in the list of default Open edX Notifications
-            msg_type = notifications_service.get_notification_type('open-edx.xblock.group-project.grades-posted')
-
-            # get the activity name which is simply our hosting
-            # Sequence's Display Name, so call out to a new xBlock
-            # runtime Service
-            courseware_info = self.get_courseware_info(self.runtime.service(self, 'courseware_parent_info'))
-            activity_name = courseware_info['activity_name']
-            activity_location = courseware_info['activity_location']
-
-            msg = NotificationMessage(
-                msg_type=msg_type,
-                namespace=unicode(self.course_id),
-                payload={
-                    '_schema_version': 1,
-                    'activity_name': activity_name,
-                }
-            )
-
-            #
-            # add in all the context parameters we'll need to
-            # generate a URL back to the website that will
-            # present the new course announcement
-            #
-            # IMPORTANT: This can be changed to msg.add_click_link() if we
-            # have a particular URL that we wish to use. In the initial use case,
-            # we need to make the link point to a different front end website
-            # so we need to resolve these links at dispatch time
-            #
-            msg.add_click_link_params({
-                'course_id': unicode(self.course_id),
-                'activity_location': unicode(activity_location) if activity_location else '',
-            })
-
-            # Bulk publish to the 'group_project_workgroup' user scope
-            notifications_service.bulk_publish_notification_to_scope(
-                'group_project_workgroup',
-                {
-                    # I think self.workgroup['id'] is a string version of an integer
-                    'workgroup_id': group_id,
-                },
-                msg
-            )
-        except Exception, ex:  # pylint: disable=broad-except
-            # While we *should* send notification, if there is some
-            # error here, we don't want to blow the whole thing up.
-            # So log it and continue....
-            log.exception(ex)
-
-    def _get_stage_timer_name(self, stage, timer_name_suffix):
-        return '{location}-{stage}-{timer_name_suffix}'.format(
-            location=self.location,
-            stage=stage.id,
-            timer_name_suffix=timer_name_suffix
-        )
-
-    def _set_activity_timed_notification(self, course_id, msg_type, stage, activity_date, send_at_date,
-                                         services, timer_name_suffix):
-
-        stage_name = stage.name
-        notifications_service = services.get('notifications')
-        courseware_parent_info = services.get('courseware_parent_info')
-
-        courseware_info = self.get_courseware_info(courseware_parent_info)
-
-        activity_name = courseware_info['activity_name']
-        activity_location = courseware_info['activity_location']
-
-        project_location = courseware_info['project_location']
-
-        activity_date_tz = activity_date.replace(tzinfo=pytz.UTC)
-        send_at_date_tz = send_at_date.replace(tzinfo=pytz.UTC)
-
-        msg = NotificationMessage(
-            msg_type=notifications_service.get_notification_type(msg_type),
-            namespace=unicode(course_id),
-            payload={
-                '_schema_version': 1,
-                'activity_name': activity_name,
-                'stage': stage_name,
-                'due_date': activity_date_tz.strftime('%-m/%-d/%-y'),
-            }
-        )
-
-        #
-        # add in all the context parameters we'll need to
-        # generate a URL back to the website that will
-        # present the new course announcement
-        #
-        # IMPORTANT: This can be changed to msg.add_click_link() if we
-        # have a particular URL that we wish to use. In the initial use case,
-        # we need to make the link point to a different front end website
-        # so we need to resolve these links at dispatch time
-        #
-        msg.add_click_link_params({
-            'course_id': unicode(course_id),
-            'activity_location': unicode(activity_location),
-        })
-
-        notifications_service.publish_timed_notification(
-            msg=msg,
-            send_at=send_at_date_tz,
-            # send to all students participating in this project
-            scope_name='group_project_participants',
-            scope_context={
-                'course_id': unicode(course_id),
-                'content_id': unicode(project_location),
-            },
-            timer_name=self._get_stage_timer_name(stage, timer_name_suffix),
-            ignore_if_past_due=True  # don't send if we're already late!
-        )
-
-    def on_studio_published(self, course_id, services):
-        """
-        A hook into when this xblock is published in Studio. When we are published we should
-        register a Notification to be send on key dates
-        """
-        try:
-            log.info('GroupActivityXBlock.on_published() on location = {}'.format(self.location))
-
-            # see if we are running in an environment which has Notifications enabled
-            notifications_service = services.get('notifications')
-            if notifications_service:
-                # set (or update) Notification timed message based on
-                # the current key dates
-                for stage in self.stages:
-
-                    # if the stage has a opening date, then send a msg then
-                    if stage.open_date:
-                        self._set_activity_timed_notification(
-                            course_id,
-                            u'open-edx.xblock.group-project.stage-open',
-                            stage,
-                            datetime.combine(stage.open_date, datetime.min.time()),
-                            datetime.combine(stage.open_date, datetime.min.time()),
-                            services,
-                            'open'
-                        )
-
-                    # if the stage has a close date, then send a msg then
-                    if stage.close_date:
-                        self._set_activity_timed_notification(
-                            course_id,
-                            u'open-edx.xblock.group-project.stage-due',
-                            stage,
-                            datetime.combine(stage.close_date, datetime.min.time()),
-                            datetime.combine(stage.close_date, datetime.min.time()),
-                            services,
-                            'due'
-                        )
-
-                        # and also send a notice 3 days earlier
-                        self._set_activity_timed_notification(
-                            course_id,
-                            u'open-edx.xblock.group-project.stage-due',
-                            stage,
-                            datetime.combine(stage.close_date, datetime.min.time()),
-                            datetime.combine(stage.close_date, datetime.min.time()) - timedelta(days=3),
-                            services,
-                            'coming-due'
-                        )
-
-        except Exception, ex:  # pylint: disable=broad-except
-            log.exception(ex)
-
-    def on_before_studio_delete(self, course_id, services):  # pylint: disable=unused-argument
-        """
-        A hook into when this xblock is deleted in Studio, for xblocks to do any lifecycle
-        management
-        """
-        log.info('GroupActivityXBlock.on_before_delete() on location = {}'.format(self.location))
-
-        try:
-            # see if we are running in an environment which has Notifications enabled
-            notifications_service = services.get('notifications')
-            if notifications_service:
-                # If we are being delete, then we should remove any NotificationTimers that
-                # may have been registered before
-                for stage in self.stages:
-                    notifications_service.cancel_timed_notification(
-                        self._get_stage_timer_name(stage, 'open')
-                    )
-
-                    notifications_service.cancel_timed_notification(
-                        self._get_stage_timer_name(stage, 'due')
-                    )
-
-                    notifications_service.cancel_timed_notification(
-                        self._get_stage_timer_name(stage, 'coming-due')
-                    )
-
-        except Exception, ex:  # pylint: disable=broad-except
-            log.exception(ex)
